Update Sync sites budges
